chore: updates example to apollo server 4 and remove old example
