--- examples/dev/hello/hello.js
@@ -1,30 +1,0 @@
-const { ApolloServer, gql } = require('apollo-server');
-const { buildSubgraphSchema } = require("@apollo/subgraph")
-
-// The GraphQL schema
-const typeDefs = gql`
-  type Query {
-    "A simple type for getting started!"
-    hello: String
-  }
-`;
-
-// A map of functions which return data for the schema.
-const resolvers = {
-  Query: {
-    hello: () => 'world',
-    goodbye: () => 'everybody'
-  },
-};
-
-const server = new ApolloServer({
-  schema: buildSubgraphSchema({
-    typeDefs,
-    resolvers,
-  })
-});
-
-server.listen({ port: 4001 }).then(({ url }) => {
-  console.log(`🚀 Server ready at ${url}`);
-});
-
