--- examples/dev/hello/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "apollo-subgraph",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node hello.js",
-    "dev": "concurrently \"npm run start\" \"cargo rover dev --name 'hello' --url 'http://localhost:4001'\""
-  },
-  "keywords": [],
-  "author": "Apollo Developers <opensource@apollographql.com>",
-  "license": "MIT",
-  "dependencies": {
-    "@apollo/subgraph": "^2.0.5",
-    "apollo-server": "^3.10.0",
-    "graphql": "^16.5.0"
-  },
-  "devDependencies": {
-    "concurrently": "7.6.0"
-  }
-}
