--- examples/dev/reloader/reloader.js
@@ -1,27 +1,0 @@
-const { buildSubgraphSchema } = require('@apollo/subgraph');
-
-const { ApolloServer, gql } = require('apollo-server');
-
-// The GraphQL schema
-const typeDefs = gql`
-  type Query {
-    "A simple type for getting started!"
-    me: String
-  }
-`;
-
-const resolvers = {
-  Query: {
-    me() {
-      return "@mara"
-    }
-  }
-};
-
-const server = new ApolloServer({
-  schema: buildSubgraphSchema({ typeDefs, resolvers })
-});
-
-server.listen({ port: 4004 }).then(({ url }) => {
-    console.log(`🚀 Server ready at ${url}`);
-});
