+++ examples/supergraph-demo/products/package.json
@@ -1,18 +1,20 @@
 {
-  "name": "reloader",
+  "name": "products",
   "version": "1.0.0",
   "description": "",
-  "main": "products.js",
+  "main": "products.mjs",
   "scripts": {
-    "start": "nodemon -e js,graphql",
-    "dev": "concurrently \"npm run start\" \"cargo rover dev --name 'products' --url 'http://localhost:4002'\""
+    "start": "nodemon -e mjs,graphql",
+    "dev": "cargo rover dev --name products --url http://localhost:4002",
+    "clean": "rm -rf node_modules package-lock.json"
   },
   "keywords": [],
   "author": "Apollo Developers <opensource@apollographql.com>",
   "license": "MIT",
   "dependencies": {
-    "@apollo/subgraph": "^2.0.5",
-    "apollo-server": "^3.10.0"
+    "@apollo/subgraph": "^2",
+    "@apollo/server": "^4",
+    "graphql-tag": "^2"
   },
   "devDependencies": {
     "concurrently": "7.6.0",
